Adding brand contorller

+++ MarketApi/Services/ProductsService.cs
@@ -1,19 +1,35 @@
 ﻿using MarketApi.Data;
 using MarketApi.Domain;
 using MarketApi.Models;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
 
 namespace MarketApi.Services
 {
 	public class ProductsService : IProductsService
 	{
 		private readonly IProductsRepository _productsRepository;
-		public ProductsService(IProductsRepository productsRepository)
+		private readonly IValidationService _validationService;
+		private readonly IImagesService<ProductModel> _imagesService;
+		public ProductsService(IProductsRepository productsRepository, IValidationService validationService, IImagesService<ProductModel> imagesService)
 		{
 			_productsRepository = productsRepository;
+			_validationService = validationService;
+			_imagesService = imagesService;
 		}
 
 		public async Task<ProductModel> Create(ProductModel model)
 		{
+
+			if (!await _validationService.IsValidBrandId(model.BrandId))
+			{
+				throw new Exception("Invalid BrandId number!");
+			}
+
+			var imageName = model.Image.FileName;
+
+			await _imagesService.SaveImage(model);
+			
+
 			var product = new Product
 			{
 				Id = model.Id,
@@ -21,7 +37,9 @@ namespace MarketApi.Services
 				Description = model.Description,
 				Price = model.Price,
 				Quantity = model.Quantity,
-				OutOfStock = model.Quantity <= 0
+				IsInStock = model.Quantity > 0,
+				BrandId = model.BrandId,
+				ImageUrl = "ProductImages/" + imageName
 			};
 
 			var createdProduct = await _productsRepository.Create(product);
@@ -33,7 +51,9 @@ namespace MarketApi.Services
 				Description = createdProduct.Description,
 				Price = createdProduct.Price,
 				Quantity = createdProduct.Quantity,
-				OutOfStock = createdProduct.OutOfStock
+				IsInStock = createdProduct.IsInStock,
+				BrandId = createdProduct.BrandId,
+				ImageUrl = createdProduct.ImageUrl
 			};
 
 			return result;
@@ -42,6 +62,10 @@ namespace MarketApi.Services
 
 		public async Task<bool> Delete(int id)
 		{
+			if(!await _imagesService.DeleteImage(id))
+			{
+				return false;
+			}
 			return await _productsRepository.Delete(id);
 		}
 
@@ -52,12 +76,14 @@ namespace MarketApi.Services
 			{
 				var model = new ProductModel
 				{
-					Id = productFromDb.Id,															// FIX
+					Id = productFromDb.Id,                                                          // FIX
 					Name = productFromDb.Name,
 					Description = productFromDb.Description,
 					Price = productFromDb.Price,
 					Quantity = productFromDb.Quantity,
-					OutOfStock = productFromDb.OutOfStock
+					IsInStock = productFromDb.IsInStock,
+					BrandId = productFromDb.BrandId,
+					ImageUrl = productFromDb.ImageUrl
 				};
 				return model;
 			}
@@ -77,7 +103,10 @@ namespace MarketApi.Services
 					Description = product.Description,
 					Price = product.Price,
 					Quantity = product.Quantity,
-					OutOfStock = product.OutOfStock
+					IsInStock = product.IsInStock,
+
+					BrandId = product.BrandId,
+					ImageUrl = product.ImageUrl
 				};
 				models.Add(model);
 			}
@@ -87,6 +116,10 @@ namespace MarketApi.Services
 
 		public async Task<ProductModel> Update(int id, ProductModel model)
 		{
+			
+
+			await _imagesService.UpdateImage(id, model);
+
 			var product = new Product
 			{
 				Id = id,
@@ -94,8 +127,11 @@ namespace MarketApi.Services
 				Description = model.Description,
 				Price = model.Price,
 				Quantity = model.Quantity,
-				OutOfStock = model.Quantity <= 0
-			};
+				IsInStock = model.Quantity > 0,
+				BrandId = model.BrandId,
+				ImageUrl = "ProductImages/" + model.Image.FileName
+		};
+			
 			var updatedProduct = await _productsRepository.Update(id, product);
 			var result = new ProductModel
 			{
@@ -104,9 +140,89 @@ namespace MarketApi.Services
 				Description = updatedProduct.Description,
 				Price = updatedProduct.Price,
 				Quantity = updatedProduct.Quantity,
-				OutOfStock = updatedProduct.OutOfStock
+				IsInStock = updatedProduct.IsInStock,
+
+				BrandId = updatedProduct.BrandId,
+				ImageUrl = updatedProduct.ImageUrl
 			};
 			return result;
+		}
+
+		public async Task<IEnumerable<ProductModel>> GetBySearchTerm(string searchTerm)
+		{
+			var productsFromDb = await _productsRepository.GetProductsBySearchTerm(searchTerm);
+			var models = new List<ProductModel>();
+
+			foreach (var product in productsFromDb)
+			{
+				var model = new ProductModel
+				{
+					Id = product.Id,
+					Name = product.Name,
+					Description = product.Description,
+					Price = product.Price,
+					Quantity = product.Quantity,
+					IsInStock = product.IsInStock,
+
+					BrandId = product.BrandId,
+					ImageUrl = product.ImageUrl
+				};
+				models.Add(model);
+			}
+
+			return models;
+		}
+
+		public async Task<IEnumerable<ProductModel>> FilterProducts(double minPrice, double maxPrice)
+		{
+			var models = new List<ProductModel>();
+
+			var productsFromDb = await _productsRepository.FilterProducts(minPrice, maxPrice);
+
+			foreach (var product in productsFromDb)
+			{
+				var model = new ProductModel
+				{
+					Id = product.Id,
+					Name = product.Name,
+					Description = product.Description,
+					Price = product.Price,
+					Quantity = product.Quantity,
+					IsInStock = product.IsInStock,
+
+					BrandId = product.BrandId,
+					ImageUrl = product.ImageUrl
+				};
+				models.Add(model);
+			}
+
+			return models;
+
+		}
+
+		public async Task<IEnumerable<ProductModel>> GetProductsOrderBy(string orderBy)
+		{
+			var models = new List<ProductModel>();
+			var productsFromDb = await _productsRepository.GetProductsOrderBy(orderBy);
+
+			foreach (var product in productsFromDb)
+			{
+				var model = new ProductModel
+				{
+					Id = product.Id,
+					Name = product.Name,
+					Description = product.Description,
+					Price = product.Price,
+					Quantity = product.Quantity,
+					IsInStock = product.IsInStock,
+
+					BrandId = product.BrandId,
+					ImageUrl = product.ImageUrl
+				};
+				models.Add(model);
+			}
+
+			return models;
 		}
 	}
 }
